+++ models/contacts.json
@@ -58,11 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "lYKa8_BctsmjqKD_VYmGZ",
-    "name": "andrii",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "222123123"
   }
 ]
